refactor: inline trivial gateway/auth index re-exports
Remove three single-purpose barrel files (chatgpt/, gemini/, settings/) and
update importers to point at the underlying modules directly.
