--- packages/server/src/gateway/auth/chatgpt/index.ts
@@ -1,1 +1,0 @@
-export { ChatGPTOAuthModule } from "./chatgpt-oauth-module.js";
