--- packages/server/src/gateway/auth/gemini/index.ts
@@ -1,1 +1,0 @@
-export { GeminiCliModule } from "./cli-module.js";
