--- packages/server/src/gateway/auth/settings/index.ts
@@ -1,5 +1,0 @@
-export {
-  type AgentSettings,
-  AgentSettingsStore,
-} from "./agent-settings-store.js";
-export { AuthProfilesManager } from "./auth-profiles-manager.js";
