+++ packages/server/src/gateway/platform.ts
@@ -9,7 +9,7 @@ import type {
 import type { AgentMetadataStore } from "./auth/agent-metadata-store.js";
 import type { McpProxy } from "./auth/mcp/proxy.js";
 import type { ProviderOAuthStateStore } from "./auth/oauth/state-store.js";
-import type { AgentSettingsStore } from "./auth/settings/index.js";
+import type { AgentSettingsStore } from "./auth/settings/agent-settings-store.js";
 import type { ModelPreferenceStore } from "./auth/settings/model-preference-store.js";
 import type { UserAgentsStore } from "./auth/user-agents-store.js";
 import type { ChannelBindingService } from "./channels/index.js";
