+++ packages/server/src/gateway/routes/public/agents.ts
@@ -14,7 +14,7 @@ import type { AgentMetadataStore } from "../../auth/agent-metadata-store.js";
 import type {
   AgentSettings,
   AgentSettingsStore,
-} from "../../auth/settings/index.js";
+} from "../../auth/settings/agent-settings-store.js";
 import { buildDefaultSettingsFromSource } from "../../auth/settings/template-utils.js";
 import type { SettingsTokenPayload } from "../../auth/settings/token-service.js";
 import type { UserAgentsStore } from "../../auth/user-agents-store.js";
